Fix filter param type, move types to planning.types
